requirements.txt
streamlit==1.35.0
langchain==0.1.17
langchain-openai==0.1.6
langchain-core==0.1.50
langchain-community==0.0.27
openai==1.30.1
python-dotenv==1.0.1
fpdf==1.7.2
PyPDF2==3.0.1
